Add alternative bindings
When calling Augeas::create, we return bindings that behave more
Ruby-ish. To this end, we set up a class Augeas::Facade, an instance of
which is returned by Augeas::create

The C bindings in _augeas.c have a few methods to undo some of the return
value munging that the old bindings do so that the Ruby code for
Augeas::Facade can easily create targeted exceptions reflecting the error,
based on the result from aug_error.

The tests in tests/tc_facade.rb are the same that were used for the augeas
gem, maintained in the fork at https://github.com/dotdoom/augeas. They pass
which means that Augeas::Facade is a complete replacement for the bindings
in that repo.