Statically link `libaugeas` to remove the runtime dependency
Currently this gem has a compile-time dependency on `libaugeas0-dev` and a run-time dependency on `libaugeas0`. We are trying to avoid compiling `ruby-augeas` on our production hosts and so we are looking to distribute the `ruby-augeas` gem bundled with the pre-compiled native extension. By statically linking `augeas` and `libxml2`, we end up with a gem that can be used without any additional dependencies needing to be installed. The approach here was largely inspired by `nokogiri`, who seem to do something very similar (although they also allow disabling static compilation with `--disable-static`).